+++ ChunkyMode/HotCompilerNamespace/HotCompilerMain.cs
@@ -24,9 +24,9 @@ namespace HotCompilerNamespace
             }
 
             {
-                var methodToReload = typeof(FireLaser).GetMethod(nameof(FireLaser.OnEnter), allFlags);
-                var newMethod = typeof(HotReloadMain).GetMethod(nameof(ModifiedGolemFireLaserOnEnter), allFlags);
-                new Hook(methodToReload, newMethod);
+                //var methodToReload = typeof(HDeMods.ChunkySimulacrum).GetMethod(nameof(HDeMods.ChunkySimulacrum.InfiniteTowerWaveController_Initialize), allFlags);
+                var newMethod = typeof(HotReloadMain).GetMethod(nameof(InfiniteTowerWaveController_Initialize_Overide), allFlags);
+                //new Hook(methodToReload, newMethod);
             }
         }
 
@@ -41,7 +41,19 @@ namespace HotCompilerNamespace
         {
             var cursor = new ILCursor(il);
             cursor.Emit(OpCodes.Ldarg_0);
-            cursor.Emit(OpCodes.Call, typeof(BaseState).GetMethod(nameof(BaseState.OnEnter), allFlags));
+            //cursor.Emit(OpCodes.Call, typeof(HDeMods.ChunkySimulacrum).GetMethod(nameof(HDeMods.ChunkySimulacrum.InfiniteTowerWaveController_Initialize), allFlags));
+        }
+
+        private static void InfiniteTowerWaveController_Initialize_Overide(ILContext il) {
+            ILCursor c = new ILCursor(il);
+            if (!c.TryGotoNext(moveType: MoveType.After,
+                    x => x.MatchLdfld<Run>("difficultyCoefficient"))
+               ) {
+                HDeMods.Log.Fatal("Failed to hook into InfiniteTowerWaveController.Initialize!");
+                return;
+            }
+
+            c.EmitDelegate<RuntimeILReferenceBag.FastDelegateInvokers.Func<float, float>>(b => b * 1.1f);
         }
 
         private static void ModifiedGolemFireLaserOnEnter(FireLaser self)